feat!: Initial custom (De-)Serializers implementation (#1860)
* Implement initial version of DeSerializers

* Fix (de-)serializers types

* Improevements after merge

* Initial state

* Apply changes for v2 request builder

* Provide realistic API without service level

* Initial discussion on services

* Update function imports

* Break everything

* Apply suggestions from code review

* Apply suggestions from code review

* Update v2

* Init v4

* Apply changes to v4 request builders

* Fix remainin transpilation errors

* Add tests for default (de-)serializers

* Fix all occurences of uriconverter

* Fix common entity and fieldbuilder types

* Remove edmToTs and tsToEdm

* Fix more tests

* tip

* Fix custom fields and batch partially

* Fix entity builder

* Fix entity.spec

* Fix key fields

* Fix more tests

* Fix filter factory

* feat: De-serializer generator part 1 (#1871)

* WIP

* WIP

* tests

* Fix some v2 parts

* Fix batch

* Fix generation

* Fix enumerability of schema

* Add nav properties

* Fix entity deserializer tests

* add generation

* run prettier

* Fix remaining v2 tests

* first adjustments type tests

* feat: De-serializer generator part 3 (#1884)

* fix collection + enum + action

* fix UTs

* ITs + type tests

* lint

* Solve circular dependencies for v2

* Fix circular dependency

* fix: filter function (deserialiser) (#1900)

* WIP

* WIP

* fix common entity

* lint

* lint

* Fix common tests

* Regenerate

* Fix v4 tests

* move all generic to service class (#1901)

* linter fix test

* add files

* Fix function imports for new name


* linter

* Restore package.json

* Fix some integration tests

* Remove unused code

* Fix test utils

* remove deserializer genric type on orderables

* half fixed type tests odata v2

* Fix complex types types

* Fix e2e tests

* Fix entity builder types

* Fix entity builder types

* before fix linter

* linter prettier

* before fix generator test

* fix tests

* linter

* linter

* adjust api

* depcheck

* depcheck

* linter

* dummy change

* remark junjie direct return

* export type

* Update edmToTs

* Regenerate

* Fix generator test

* Regenerate

* Fix eslint config

* Fix some docs

* Fix references to entityconstructor

* Improve docs on entity api params

* Fix API exports

* Improve docs

* Fix incorrect rename

* Lint

* Remove unused request builder

* Remove unused request builder

* Fix request biulder tests

* Remove unused files

+++ test-packages/integration-tests/test/odata-negative-case.spec.ts
@@ -17,8 +17,14 @@ describe('odata negative tests', () => {
     if (!existsSync(testOutputRootDir)) {
       await promises.mkdir(testOutputRootDir);
     }
-    await promises.rmdir(testDir, { recursive: true });
+    if (existsSync(testDir)) {
+      await promises.rmdir(testDir, { recursive: true });
+    }
     await promises.mkdir(testDir);
+  });
+
+  afterAll(async () => {
+    await promises.rmdir(testDir, { recursive: true });
   });
 
   it('should fail on faulty edmx', async () => {
