+++ test-packages/integration-tests/test/openapi-negative-case.spec.ts
@@ -17,8 +17,14 @@ describe('openapi negative tests', () => {
     if (!existsSync(testOutputRootDir)) {
       await promises.mkdir(testOutputRootDir);
     }
-    await promises.rmdir(testDir, { recursive: true });
+    if (existsSync(testDir)) {
+      await promises.rmdir(testDir, { recursive: true });
+    }
     await promises.mkdir(testDir);
+  });
+
+  afterAll(async () => {
+    await promises.rmdir(testDir, { recursive: true });
   });
 
   it('should fail on generation for faulty spec file', async () => {
